+++ src/Method/BanChatMemberMethod.php
@@ -10,13 +10,14 @@ use TgBotApi\BotApiBase\Method\Traits\FillFromArrayTrait;
 use TgBotApi\BotApiBase\Method\Traits\UserIdVariableTrait;
 
 /**
- * Class KickChatMemberMethod.
+ * Class BanChatMemberMethod.
  *
- * @see https://core.telegram.org/bots/api#kickchatmember
+ * Use this method to ban a user in a group, a supergroup or a channel. Returns True on success.
+ * Replaces the method previously known as kickChatMember.
  *
- * @deprecated Absent from the current Bot API documentation. Use BanChatMemberMethod instead.
+ * @see https://core.telegram.org/bots/api#banchatmember
  */
-class KickChatMemberMethod implements KickMethodAliasInterface
+class BanChatMemberMethod implements KickMethodAliasInterface
 {
     use FillFromArrayTrait;
     use ChatIdVariableTrait;
@@ -32,20 +33,26 @@ class KickChatMemberMethod implements KickMethodAliasInterface
     public $untilDate;
 
     /**
-     * KickChatMemberMethod constructor.
+     * Optional. Pass True to delete all messages from the chat for the user that is being removed.
+     * If False, the user will be able to see messages in the group that were sent before the user was removed.
+     * Always True for supergroups and channels.
      *
+     * @var bool|null
+     */
+    public $revokeMessages;
+
+    /**
      * @param array|null $data
      *
      * @throws \TgBotApi\BotApiBase\Exception\BadArgumentException
-     *
      */
-    public static function create(int|string $chatId, int $userId, array $data = null): KickChatMemberMethod
+    public static function create(int|string $chatId, int $userId, ?array $data = null): BanChatMemberMethod
     {
         $static = new static();
         $static->chatId = $chatId;
         $static->userId = $userId;
         if ($data) {
-            $static->fill(data: $data);
+            $static->fill($data);
         }
 
         return $static;