+++ src/Traits/GetMethodTrait.php
@@ -7,6 +7,7 @@ namespace TgBotApi\BotApiBase\Traits;
 use TgBotApi\BotApiBase\Exception\ResponseException;
 use TgBotApi\BotApiBase\Method\GetChatAdministratorsMethod;
 use TgBotApi\BotApiBase\Method\GetChatMemberMethod;
+use TgBotApi\BotApiBase\Method\GetChatMemberCountMethod;
 use TgBotApi\BotApiBase\Method\GetChatMembersCountMethod;
 use TgBotApi\BotApiBase\Method\GetChatMethod;
 use TgBotApi\BotApiBase\Method\GetChatMenuButtonMethod;
@@ -88,6 +89,17 @@ trait GetMethodTrait
 
     /**
      * @throws ResponseException
+     */
+    public function getChatMemberCount(GetChatMemberCountMethod $getChatMemberCountMethod): int
+    {
+        return $this->call(method: $getChatMemberCountMethod);
+    }
+
+    /**
+     * @throws ResponseException
+     *
+     * @deprecated Telegram renamed this method to "getChatMemberCount" in the Bot API; it is absent from the
+     *             current documentation. Use getChatMemberCount() instead.
      */
     public function getChatMembersCount(GetChatMembersCountMethod $getChatMembersCountMethod): int
     {